requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
click==8.1.7
datetime2==0.9.4
ecdsa==0.18.0
fastapi==0.110.0
h11==0.14.0
httptools==0.6.1
idna==3.6
pyasn1==0.5.1
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0